requirements.txt
asgiref==3.3.1
Django==3.1.5
django-widget-tweaks==1.4.8
gunicorn==20.0.4
mysqlclient==2.0.3
pytz==2020.5
sqlparse==0.4.1
django-allauth~=0.44.0
requests~=2.25.1
beautifulsoup4~=4.9.3

djangorestframework~=3.12.4